pyproject.toml
[project]
name = "proiectatestat"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pillow>=12.0.0",
    "streamlit>=1.52.1",
    "watchdog>=6.0.0",
]
